requirements.txt
matplotlib~=3.7.1
numpy~=1.24.1

torch~=2.0.1+cu118
torchvision~=0.15.2+cu118

dicom2jpg~=0.1.10
Pillow~=9.3.0